fertigation-backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv
pydantic
pandas
numpy
scikit-learn
joblib
